add 'money' type
Hi. Today I used your class while I was working on a project, so I noticed that the type 'money', which is defined in sql, cannot be used.
It is considered as double, but then it will give error because of the money sign. So I added an if and I moved the break statement, in order to condider the 'money' type
as a string. Let me know what you think about this change.